+++ front-end/src/Components/CardSlides/CardSlides.jsx
@@ -162,20 +162,45 @@ export default function CardSlides({
   };
 
   const handleToDelete = () => {
-    axios
-      .delete(`http://localhost:3000/admin/slide-delete/${id_type}`, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      })
-      .then((response) => {
-        console.log("Exam deleted successfully", response.data);
-        onDelete();
-      })
-      .catch((error) => {
-        console.error("Error deleting exam", error);
-      });
+    Swal.fire({
+      title: "هل أنت متأكد من حذف الملف",
+      text: "لا يمكنك التراجع عن هذه العملية!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "نعم، احذفها",
+      cancelButtonText: "إلغاء",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios
+          .delete(`http://localhost:3000/admin/slide-delete/${id_type}`, {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          })
+          .then((response) => {
+            console.log("Slide deleted successfully", response.data);
+            onDelete();
+            Swal.fire({
+              icon: "success",
+              title: "تم حذف الملف بنجاح",
+              timer: 1000,
+              showConfirmButton: false,
+            });
+          })
+          .catch((error) => {
+            console.error("Error deleting slide", error);
+            Swal.fire({
+              icon: "error",
+              title: "حدث خطأ أثناء الحذف",
+              text: error.response?.data?.message || "حاول مرة أخرى لاحقًا",
+            });
+          });
+      }
+    });
   };
+
   return (
     <div className="card-slide">
       <div className="card-bg-slide">
